requirements.txt
pandas
psycopg2-binary
plotly
streamlit
requests